Docstringでコメントを書く #8
Docstringでのコメントを追加しました　#8
